jobbot/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
openpyxl>=3.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
ollama>=0.1.0
click>=8.1.0
